requirements.txt
numpy
networkx
qubovert
dimod
dwave_qbsolv
dwave-neal
dwave-tabu
pytest
hypothesis
hypothesis_networkx